# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Therapist API's",
        "description": "API's for the Therapist project",
        "contact": {
            "email": "gursimrat1989@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/therapist-dir/public/api",
    "paths": {
...
        "/issues/list": {
            "get": {
                "tags": [
                    "Issues"
                ],
                "summary": "Issues List",
                "operationId": "listIssues",
                "parameters": [
                    {
                        "name": "X-custom-header",
                        "in": "header",
                        "description": "header",
                        "required": true,
                        "type": "string",
                        "default": "7890abcdefghijkl"
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/issue/add": {
            "post": {
                "tags": [
                    "Issues"
                ],
                "summary": "Add issue",
                "operationId": "addIssue",
                "parameters": [
                    {
                        "name": "X-custom-header",
                        "in": "header",
                        "description": "header",
                        "required": true,
                        "type": "string",
                        "default": "7890abcdefghijkl"
                    },
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/issues/add/description": {
            "post": {
                "tags": [
                    "Issues"
                ],
                "summary": "Add issue description by users",
                "operationId": "issuesDescriptionAdd",
                "parameters": [
                    {
                        "name": "X-custom-header",
                        "in": "header",
                        "description": "header",
                        "required": true,
                        "type": "string",
                        "default": "7890abcdefghijkl"
                    },
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "issue_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "issue_description",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/issues/list/description": {
            "get": {
                "tags": [
                    "Issues"
                ],
                "summary": "List issue description by users",
                "operationId": "issuesDescriptionList",
                "parameters": [
                    {
                        "name": "X-custom-header",
                        "in": "header",
                        "description": "header",
                        "required": true,
                        "type": "string",
                        "default": "7890abcdefghijkl"
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "issue_id",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/issues/get/name": {
            "get": {
                "tags": [
                    "Issues"
                ],
                "summary": "Get issue name by id",
                "operationId": "issuesGetName",
                "parameters": [
                    {
                        "name": "X-custom-header",
                        "in": "header",
                        "description": "header",
                        "required": true,
                        "type": "string",
                        "default": "7890abcdefghijkl"
                    },
                    {
                        "name": "issue_id",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/issues/description/update": {
            "post": {
                "tags": [
                    "Issues"
                ],
                "summary": "Update issue description or explanation",
                "operationId": "issuesDescriptionUpdate",
                "parameters": [
                    {
                        "name": "X-custom-header",
                        "in": "header",
                        "description": "header",
                        "required": true,
                        "type": "string",
                        "default": "7890abcdefghijkl"
                    },
                    {
                        "name": "description_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/issues/description/delete": {
            "post": {
                "tags": [
                    "Issues"
                ],
                "summary": "Delete issue description or explanation",
                "operationId": "issuesDescriptionDelete",
                "parameters": [
                    {
                        "name": "X-custom-header",
                        "in": "header",
                        "description": "header",
                        "required": true,
                        "type": "string",
                        "default": "7890abcdefghijkl"
                    },
                    {
                        "name": "description_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/issues/search": {
            "get": {
                "tags": [
                    "Issues"
                ],
                "summary": "Insurance Search",
                "operationId": "searchIssues",
                "parameters": [
                    {
                        "name": "X-custom-header",
                        "in": "header",
                        "description": "header",
                        "required": true,
                        "type": "string",
                        "default": "7890abcdefghijkl"
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "keyword",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/issue/delete": {
            "get": {
                "tags": [
                    "Issues"
                ],
                "summary": "Delete issue",
                "operationId": "deleteIssue",
                "parameters": [
                    {
                        "name": "X-custom-header",
                        "in": "header",
                        "description": "header",
                        "required": true,
                        "type": "string",
                        "default": "7890abcdefghijkl"
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "issue_id",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
...
}